requirement.txt
python-dotenv
pinecone
git+https://github.com/openai/whisper.git@main#egg=openai-whisper
gTTS
moviepy
langchain
langchain_google_genai
langchain_pinecone
torch
torchaudio